+++ source/client/editor/cardBuilder/cardSubBuilders/CardArtwork.tsx
@@ -1,6 +1,6 @@
 import { ICard } from 'client/editor/card/card-interfaces';
 import { ToolkitComponent, IToolkitComponentProps, IToolkitComponentState, TDidUpdateSnapshot } from 'mn-toolkit';
-import { getCroppedArtworkBase64 } from 'mn-tools';
+import { getCroppedArtworkBase64, preloadImage } from 'mn-tools';
 
 interface ICardArtworkProps extends IToolkitComponentProps {
   card: ICard;
@@ -89,7 +89,7 @@ export class CardArtwork extends ToolkitComponent<ICardArtworkProps, ICardArtwor
     if (this.state.loadArtwork) {
       requestAnimationFrame(() => requestAnimationFrame(() => app.$errorManager.handlePromise(this.loadArtwork())));
     } else {
-      this.props.onReady();
+      requestAnimationFrame(() => requestAnimationFrame(() => this.props.onReady()));
     }
   }
 
@@ -125,6 +125,7 @@ export class CardArtwork extends ToolkitComponent<ICardArtworkProps, ICardArtwor
       artworkData = artworkBg;
     }
 
+    await preloadImage(artworkData);
     this.setState({ loadArtwork: false, artworkData });
   }